+++ benefits/src/lib/config.ts
@@ -55,7 +55,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = Object.freeze({
     language: "typescript",
     openapiDocVersion: "1.0.0",
-    sdkVersion: "2.0.0",
-    genVersion: "2.272.1",
-    userAgent: "speakeasy-sdk/typescript 2.0.0 2.272.1 1.0.0 @wingspan/benefits",
+    sdkVersion: "2.0.1",
+    genVersion: "2.277.0",
+    userAgent: "speakeasy-sdk/typescript 2.0.1 2.277.0 1.0.0 @wingspan/benefits",
 });